requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
httpx>=0.25.2
PyMuPDF>=1.23.0
python-docx>=1.0.0
sentence-transformers>=2.2.0
pinecone-client>=3.0.0
faiss-cpu>=1.7.0
google-generativeai>=0.8.5
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
redis>=5.0.0
nltk>=3.8
numpy>=1.24.0
scikit-learn>=1.3.0
# Additional packages for the fixed document processor
lxml>=4.9.0
beautifulsoup4>=4.12.0
chardet>=5.2.0
mail-parser>=1.7.1